+++ frontend/src/metabase/admin/performance/components/StrategyForm.tsx
@@ -41,11 +41,14 @@ import type {
 } from "metabase-types/api";
 import { CacheDurationUnit } from "metabase-types/api";
 
-import { strategyValidationSchema } from "../constants/complex";
 import { defaultCronSchedule, rootId } from "../constants/simple";
 import { useIsFormPending } from "../hooks/useIsFormPending";
 import { isModelWithClearableCache } from "../types";
-import { getDefaultValueForField, getLabelString } from "../utils";
+import {
+  getDefaultValueForField,
+  getLabelString,
+  strategyValidationSchema,
+} from "../utils";
 
 import PerformanceAppStyles from "./PerformanceApp.module.css";
 import S from "./StrategyForm.module.css";
