+++ frontend/src/metabase/admin/performance/constants/complex.ts
@@ -5,7 +5,6 @@ import type { AdminPath } from "metabase/redux/store";
 import type { CacheableModel } from "metabase-types/api";
 
 import { PerformanceTabId, type StrategyData } from "../types";
-import { getStrategyValidationSchema, isValidStrategyName } from "../utils";
 
 import { defaultMinDurationMs } from "./simple";
 
@@ -35,40 +34,6 @@ export const getAdaptiveStrategyValidationSchema = () => {
     multiplier: positiveInteger.default(10),
   });
 };
-
-export const strategyValidationSchema = Yup.object().test(
-  "strategy-validation",
-  "The object must match one of the strategy validation schemas",
-  function (value) {
-    if (!value) {
-      return this.createError({
-        message: "Strategy is falsy",
-      });
-    }
-    const { type } = value as unknown as { type: string };
-    if (!isValidStrategyName(type)) {
-      return this.createError({
-        message: `"${type}" is not a valid strategy name`,
-        path: "type",
-      });
-    }
-    const schema = getStrategyValidationSchema(strategies[type]);
-    try {
-      schema.validateSync(value);
-      return true;
-    } catch (error: unknown) {
-      if (error instanceof Yup.ValidationError) {
-        return this.createError({
-          message: error.message,
-          path: error.path,
-        });
-      } else {
-        console.error("Unhandled error:", error);
-        return false;
-      }
-    }
-  },
-) as Yup.AnySchema;
 
 export const strategies = {
   inherit: {
