+++ frontend/src/metabase/admin/utils.ts
@@ -2,11 +2,11 @@ import { useLayoutEffect } from "react";
 import { push, replace, routerActions } from "react-router-redux";
 import { connectedReduxRedirect } from "redux-auth-wrapper/history3/redirect";
 
-import { getAdminPaths } from "metabase/admin/app/selectors";
 import { shouldShowTenantsUpsell } from "metabase/admin/people/selectors";
 import { connect, metabaseReduxContext } from "metabase/redux";
 import type { State } from "metabase/redux/store";
 import type { AdminPath } from "metabase/redux/store/admin";
+import { getAdminPaths } from "metabase/selectors/admin";
 import { getSetting } from "metabase/selectors/settings";
 
 export const createAdminRouteGuard = (routeKey: string) => {
