+++ frontend/src/metabase/route-guards.tsx
@@ -1,11 +1,11 @@
 import { routerActions } from "react-router-redux";
 import { connectedReduxRedirect } from "redux-auth-wrapper/history3/redirect";
 
-import { getAdminPaths } from "metabase/admin/app/selectors";
 import { canAccessDataStudio } from "metabase/data-studio/selectors";
 import { PLUGIN_FEATURE_LEVEL_PERMISSIONS } from "metabase/plugins";
 import { metabaseReduxContext } from "metabase/redux";
 import type { State } from "metabase/redux/store";
+import { getAdminPaths } from "metabase/selectors/admin";
 import { getSetting } from "metabase/selectors/settings";
 import { isSameOrSiteUrlOrigin } from "metabase/utils/dom";
